Factor out toolbar content components
